+++ tests/PdfExportFromHtmlFileTest.php
@@ -5,7 +5,7 @@ namespace Sfneal\ViewExport\Tests;
 use Sfneal\ViewExport\Pdf\PdfExportService;
 use Sfneal\ViewExport\Tests\Traits\PdfExportValidations;
 
-class PdfExportFromHtmlPathTest extends TestCase
+class PdfExportFromHtmlFileTest extends TestCase
 {
     use PdfExportValidations;
 
@@ -18,6 +18,6 @@ class PdfExportFromHtmlPathTest extends TestCase
     {
         parent::setUp();
 
-        $this->renderer = PdfExportService::fromHtmlPath(base_path('tests/resources/html/test.html'));
+        $this->renderer = PdfExportService::fromHtmlFile(base_path('tests/resources/html/test.html'));
     }
 }